перенесла покемонов в отдельный файл
